requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.84
langchain-openai==0.3.30
langchain-teddynote==0.5.1
langchain-experimental==0.3.4
langchain-tavily==0.2.11
beautifulsoup4==4.13.4
pdfplumber==0.11.7
faiss-cpu==1.12.0
langchain-mcp-adapters==0.1.9
fastmcp==2.12.4
streamlit==1.48.1